@@ -38,8 +38,8 @@
     "@ohif/extension-default": "3.9.0-beta.14",
     "@ohif/i18n": "3.9.0-beta.14",
     "prop-types": "^15.6.2",
-    "react": "^17.0.2",
-    "react-dom": "^17.0.2",
+    "react": "^18.3.1",
+    "react-dom": "^18.3.1",
     "react-i18next": "^12.2.2",
     "react-router": "^6.8.1",
     "react-router-dom": "^6.8.1"
