@@ -41,8 +41,8 @@
     "dcmjs": "^0.29.12",
     "lodash.debounce": "^4.17.21",
     "prop-types": "^15.6.2",
-    "react": "^17.0.2",
-    "react-dom": "^17.0.2",
+    "react": "^18.3.1",
+    "react-dom": "^18.3.1",
     "webpack": "^5.50.0",
     "webpack-merge": "^5.7.3"
   },
